# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ScrollUniversity API
  description: |
    Complete API documentation for the ScrollUniversity platform - a revolutionary Christian educational platform 
    that combines divine revelation with cutting-edge technology.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - General endpoints: 100 requests per 15 minutes
    - Authentication endpoints: 5 requests per 15 minutes
    - API endpoints: 1000 requests per hour
    
    ## Response Format
    All responses follow this structure:
    ```json
    {
      "success": true,
      "data": { ... },
      "error": null
    }
    ```
  version: 1.0.0
  contact:
    name: ScrollUniversity API Support
    email: api@scrolluniversity.org
  license:
    name: Proprietary
    url: https://scrolluniversity.org/license
servers:
  - url: https://api.scrolluniversity.org/v1
    description: Production server
  - url: https://staging-api.scrolluniversity.org/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid input data"
            details:
              type: object
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [STUDENT, FACULTY, ADMIN]
        avatar:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                name:
                  type: string
                role:
                  type: string
                  enum: [STUDENT, FACULTY]
                  default: STUDENT
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      message:
                        type: string
                        example: "Verification email sent"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user and receive JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      accessToken:
                        type: string
                      refreshToken:
                        type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                
  # Course Endpoints
...
security:
  - bearerAuth: []